+++ data-migrations/2025-07-23-1216-update-question-json.sql
@@ -1,0 +1,4 @@
+# Update the default seed data question JSON to conform to @dmptool/types 1.2.0
+UPDATE `questions`
+  SET `json` = '{"meta":{"schemaVersion":"1.0"},"type":"textArea","attributes":{"cols":20,"rows":2,"asRichText":true}}'
+  WHERE `json` LIKE '%asRichText%';